image upload with formgroup completed

+++ src/app/header/header/header.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
+import { UserService } from 'src/app/user.service';
 
 @Component({
   selector: 'app-header',
@@ -7,10 +9,12 @@ import { Router } from '@angular/router';
   styleUrls: ['./header.component.css']
 })
 export class HeaderComponent implements OnInit{
-
+  
+  userData: any;
   isLoginMode = true;
+  token:any;
 
-  constructor(private router: Router){}
+  constructor(private router: Router, private userService: UserService, private toastr: ToastrService){}
 
   ngOnInit(){
     this.isLoginMode=this.checkSession();
@@ -27,7 +31,24 @@ export class HeaderComponent implements OnInit{
   }
 
   Logout(){
-    sessionStorage.clear();
-    this.router.navigate(['/login']);
+    this.token = sessionStorage.getItem('token');
+
+    const tokenPayload = {
+      token: this.token
+    };
+
+    
+    this.userService.onLogoutUser(tokenPayload).subscribe(res =>{
+      this.userData = res;
+      if(this.userData.success==true){
+        sessionStorage.clear();
+        this.toastr.success('User Logout successfully.');
+        this.isLoginMode = true;
+        this.router.navigate(['/login']);
+      }else{
+        this.toastr.error('Something Went Wrong');
+        this.isLoginMode = false;
+      }
+    }); 
   }
 }
